requirements.txt
Django==2.0.13
pytz==2019.1
django-filter==2.2.0
djangorestframework==3.9.4
Pygments==2.3.1
